streamlit_app_files/requirements.txt
streamlit
scikit-learn
xgboost
pillow
